InterviewProject/requirements.txt
flask~=2.3.3
flask-smorest
python-dotenv~=1.0.0
marshmallow~=3.20.1
sqlalchemy~=2.0.20
flask-sqlalchemy
flask-jwt-extended
passlib~=1.7.4
flask-migrate
psycopg2-binary
Jinja2~=3.1.2
alembic~=1.12.0
xhtml2pdf
flask-cors